Move tests to correct directory
